This is a work in progress.
i got a little mixed up and need to rework the classes to account for the resturant and simplicity in the interface to deal with the file IO. Then the manager just handles the restaurants. I checked this with claude and it agreed this was a better practice then adding this all into the RestaurantManager class.

Note, as part of the update AI tossed in some methods that have not been reviewed for accuracy. Some have had minor updates or were removed.

This is a work in progress.